removed user creds processing to user/credentials
