+++ lib/user/store/mysql.js
@@ -1,5 +1,6 @@
 import Mysql from '../../mysql.js'
 import Config from '../../config.js'
+import Credentials from '../credentials.js'
 import UserBase from './base.js'
 import Permission from '../../permission/index.js'
 import { mapToDbColumn } from '../../util.js'
@@ -36,7 +37,7 @@ class UserRepoMySQL extends UserBase {
       AND g.name = ?`
 
     for (const u of await Mysql.execute(query, [username, groupName])) {
-      const { valid, needsUpgrade } = await this.validPassword(
+      const { valid, needsUpgrade } = await Credentials.validPassword(
         authTry.password,
         u.password,
         authTry.username,
@@ -47,14 +48,8 @@ class UserRepoMySQL extends UserBase {
         // they simply get upgraded on a later attempt
         if (needsUpgrade) {
           try {
-            const salt = this.generateSalt()
-            const hash = await this.hashForStorage(authTry.password, salt)
-            await Mysql.execute(
-              ...Mysql.update('nt_user', `nt_user_id=${u.id}`, {
-                password: hash,
-                pass_salt: salt,
-              }),
-            )
+            const creds = await Credentials.forStorage(authTry.password)
+            await Mysql.execute(...Mysql.update('nt_user', `nt_user_id=${u.id}`, creds))
           } catch (err) {
             console.warn(`password hash upgrade failed for user ${authTry.username}`, err)
           }
@@ -86,8 +81,7 @@ class UserRepoMySQL extends UserBase {
     delete args.inherit_group_permissions
 
     if (args.password) {
-      if (!args.pass_salt) args.pass_salt = this.generateSalt()
-      args.password = await this.hashForStorage(args.password, args.pass_salt)
+      Object.assign(args, await Credentials.forStorage(args.password, args.pass_salt))
     }
 
     const userId = await Mysql.execute(...Mysql.insert(`nt_user`, mapToDbColumn(args, userDbMap)))
