+++ lib/user/store/toml.js
@@ -5,6 +5,7 @@ import { fileURLToPath } from 'node:url'
 import { parse, stringify } from 'smol-toml'
 
 import Config from '../../config.js'
+import Credentials from '../credentials.js'
 import UserBase from './base.js'
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url))
@@ -86,7 +87,7 @@ class UserRepoTOML extends UserBase {
       if (u.username !== username) continue
       if (u.deleted) continue
 
-      const { valid, needsUpgrade } = await this.validPassword(
+      const { valid, needsUpgrade } = await Credentials.validPassword(
         authTry.password,
         u.password,
         authTry.username,
@@ -97,9 +98,7 @@ class UserRepoTOML extends UserBase {
         // the user their login
         if (needsUpgrade) {
           try {
-            const salt = this.generateSalt()
-            u.pass_salt = salt
-            u.password = await this.hashForStorage(authTry.password, salt)
+            Object.assign(u, await Credentials.forStorage(authTry.password))
             await this._save(users)
           } catch (err) {
             console.warn(`password hash upgrade failed for user ${authTry.username}`, err)
@@ -170,8 +169,7 @@ class UserRepoTOML extends UserBase {
     delete args.inherit_group_permissions
 
     if (args.password) {
-      if (!args.pass_salt) args.pass_salt = this.generateSalt()
-      args.password = await this.hashForStorage(args.password, args.pass_salt)
+      Object.assign(args, await Credentials.forStorage(args.password, args.pass_salt))
     }
 
     if (inherit === false) {
