+++ lib/user/test/index.js
@@ -3,6 +3,7 @@ import { describe, it, after, before } from 'node:test'
 
 import User from '../index.js'
 import Group from '../../group/index.js'
+import Credentials from '../credentials.js'
 
 import userJson from './user.json' with { type: 'json' }
 import groupJson from '../../group/test/group.json' with { type: 'json' }
@@ -90,40 +91,40 @@ describe('user', function () {
 
   describe('validPassword', function () {
     it('auths user with plain text password', async () => {
-      const r = await User.validPassword('test', 'test', 'demo', '')
+      const r = await Credentials.validPassword('test', 'test', 'demo', '')
       assert.deepEqual(r, { valid: true, needsUpgrade: true })
     })
 
     it('auths valid self-describing PBKDF2 password', async () => {
       const salt = '(ICzAm2.QfCa6.MN'
-      const hash = await User.hashForStorage('YouGuessedIt!', salt)
-      const r = await User.validPassword('YouGuessedIt!', hash, 'unit-test', salt)
+      const hash = await Credentials.hashForStorage('YouGuessedIt!', salt)
+      const r = await Credentials.validPassword('YouGuessedIt!', hash, 'unit-test', salt)
       assert.deepEqual(r, { valid: true, needsUpgrade: false })
     })
 
     it('rejects invalid self-describing PBKDF2 password', async () => {
       const salt = '(ICzAm2.QfCa6.MN'
-      const hash = await User.hashForStorage('YouGuessedIt!', salt)
-      const r = await User.validPassword('YouMissedIt!', hash, 'unit-test', salt)
+      const hash = await Credentials.hashForStorage('YouGuessedIt!', salt)
+      const r = await Credentials.validPassword('YouMissedIt!', hash, 'unit-test', salt)
       assert.deepEqual(r, { valid: false, needsUpgrade: false })
     })
 
     it('auths valid legacy PBKDF2-5000 password', async () => {
       const salt = '(ICzAm2.QfCa6.MN'
-      const hash = await User.hashAuthPbkdf2('YouGuessedIt!', salt, 5000)
-      const r = await User.validPassword('YouGuessedIt!', hash, 'unit-test', salt)
+      const hash = await Credentials.hashAuthPbkdf2('YouGuessedIt!', salt, 5000)
+      const r = await Credentials.validPassword('YouGuessedIt!', hash, 'unit-test', salt)
       assert.deepEqual(r, { valid: true, needsUpgrade: true })
     })
 
     it('rejects invalid legacy PBKDF2-5000 password', async () => {
       const salt = '(ICzAm2.QfCa6.MN'
-      const hash = await User.hashAuthPbkdf2('YouGuessedIt!', salt, 5000)
-      const r = await User.validPassword('YouMissedIt!', hash, 'unit-test', salt)
+      const hash = await Credentials.hashAuthPbkdf2('YouGuessedIt!', salt, 5000)
+      const r = await Credentials.validPassword('YouMissedIt!', hash, 'unit-test', salt)
       assert.deepEqual(r, { valid: false, needsUpgrade: false })
     })
 
     it('auths valid SHA1 password', async () => {
-      const r = await User.validPassword(
+      const r = await Credentials.validPassword(
         'OhNoYouDont',
         '083007777a5241d01abba70c938c60d80be60027',
         'unit-test',
@@ -132,7 +133,7 @@ describe('user', function () {
     })
 
     it('rejects invalid SHA1 password', async () => {
-      const r = await User.validPassword(
+      const r = await Credentials.validPassword(
         'OhNoYouDont',
         '083007777a5241d01abba7Oc938c60d80be60027',
         'unit-test',
