+++ lib/user/test/mysql.js
@@ -4,6 +4,7 @@ import { describe, it, before, after } from 'node:test'
 
 import User from '../index.js'
 import Group from '../../group/index.js'
+import Credentials from '../credentials.js'
 
 import groupJson from '../../group/test/group.json' with { type: 'json' }
 
@@ -99,8 +100,8 @@ describe('user (mysql)', function () {
     })
 
     it('upgrades PBKDF2-5000 to self-describing format on login', async () => {
-      const legacySalt = User.generateSalt()
-      const legacyHash = await User.hashAuthPbkdf2(testPass, legacySalt, 5000)
+      const legacySalt = Credentials.generateSalt()
+      const legacyHash = await Credentials.hashAuthPbkdf2(testPass, legacySalt, 5000)
       await seedUser(legacyHash, legacySalt)
 
       assert.ok(await User.authenticate(authCreds), 'login should succeed with legacy PBKDF2')
@@ -113,8 +114,8 @@ describe('user (mysql)', function () {
     })
 
     it('does not re-hash password already in self-describing format', async () => {
-      const salt = User.generateSalt()
-      const hash = await User.hashForStorage(testPass, salt)
+      const salt = Credentials.generateSalt()
+      const hash = await Credentials.hashForStorage(testPass, salt)
       await seedUser(hash, salt)
 
       await User.authenticate(authCreds)
